+++ CHANGELOG.md
@@ -6,6 +6,10 @@
   zip file for a year worth of data.  
 - Move ISONE Monthly Asset NCPC to DuckDB
 
+# Release 2025-06-15
+- First release of qplot.  Let's see if this is as useful as I hope it will be.
+- Modify Plotly.now() to launch the default browser if file path is not specified.
+
 # Release 2025-05-27
 - Some work on HQ water data